Platform: Debugger interface.
The Debugger interface is an object that encapsulates
any debugging functionality. It is initialized at the earliest
possible point and shutdown at the latest point.

On Window, when --leak-check is given as a command-line
option and it is run within the Visual Studio debugger,
all leaked memory is printed to the output.