Metal: Initialize metal layer and render with clear color.
This commit properly sets up a View and ViewController
so that a metal layer can be created and used within
the Renderer object.

This now allows for initializing the ViewController within the
Render module instead of the Platform like it was previously.

The logic for this is taken from Apple's Create A Custom View
sample project.