Core: Rotation class.
Represents a Euler rotation through a pitch, yaw and
roll fields. Can be converted into a single rotation
matrix.

Camera class replaces the direction vector with a
Rotation field which is used to calculate the direction
the camera should be facing.

A test case has been added to ensure the validity of the
Rotation class and the rotation matrices.