Render: API for rendering to a window and issue commands.
The Renderer::Render function has been deprecated in favor
of a more robust API to dictate what window to render
to and what graphics pipeline and draw commands to use
for a particular window.

All vertex buffers are now created outside of the Renderer.
The Main module now defines the vertices to render and
the GUI module is responsible for the drawing OctaneGUI
paint commands.