GUI: Refactor how vertex data is updated and drawn.
Vertex buffer for each OctaneGUI is now stored. OnPaint
callback is only called for windows where the state for that
window is modified. The vertex and index buffers for each
window is combined into a single buffer and uploaded.
Draw calls are made for each window and properly offset
into the combined buffers.