Menu bar, menu and menu items.
The menu module is responsible for managing the menu bar
and menus and their menu items. Observers can be assigned
to each menu item so menu can be created or an action can
be invoked.

This commit begins with a 'File' and 'Help' menu bar items.
The 'Quit' File command and About Help command are the
initial items for now.