KeaAnimation component.
This component manages a list of KeaAnimationClip objects that
tween a property on another component of the owning entity
through the reflection system.

A KeaAnimcationClip describes what component and property is going
to be animated.

A KeaAnimatable trait with the 'reflect_trait' attribute is
registered for the 'Val', 'Quat', and 'f32' types for now.

KeaAnimationComplete events are emitted whenever a KeaAnimationClip
has finished.