Horizontal and vertical scrollbars for KeaScrollables.
A KeaScrollable component will spawn a child container that
holds the horizontal and vertical scrollbars. The scrollbar
position and size are updated whenever the contents of the
KeaScrollable children changes. The scrollbars can be dragged
with the mouse and the vertical scrollbar can be moved with
the mouse wheel.

The bars themselves will appear small at the edges when the
mouse is not near them. When near, they will play an animation
to expand to allow for interaction.

A new module 'scroll' has been introduced in the 'controls' module
to hold all the scroll components and systems.