KeaExpander refactor to allow for custom header.
The previous version only allowed creating a bundle with a header
provided. Custom headers now allow more flexibility on how an
expander can look and behave.

KeaExpanderEvent was also added to notify when the state of
the expander changes.

The refactor includes moving the expander file into a folder
module.