Create 'lstalk' crate for language server communication.
The LSPService object manages the communication between the LevelSketch
app and language servers. This is currently the base implementation
of the protocol which only initializes the communication between
the service and the language server.

The communication is handled on a separate thread and utilizes the
'async-process' crate for polling the request/response cycle
asynchronously.