Refactor language service and server.
This refactor moves the owner of the thread to be in the language
server object. Previously, this lived in the service. This move
allows the service to host multiple servers.

This also adds support for communicating events from the language
server thread to the main thread.

Error enums no longer implement the Display trait. Error messages
should be handled by the caller.