Make request for semantic tokens of a document symbol request.
The initialize response now parses and stores the language
server's capabilities. These capabilities contains the semantic
token types and modifiers needed for resolving responses from
semantic tokens request for a document.

The responses for the document symbols and semantic tokens are now
forwarded to the LanguageServerRunner. All resolves will be
performed in this object.