KeaPanelCloseBehavior and KeaPanelClose event.
The close behavior defines how the panel responds to a user
clicking the close button. The default behavior is to despawn
the panel. This can be changed to emit an event instead which
allows custom logic to be performed instead of despawning.

The Tools context menu will now toggle the visibility of the panel
when selected. The panels themselves will now hide instead of
despawning.