Refactor KeaTree bundle to make KeaTree the root.
Previously, the KeaTree component was added as the parent of the
actual list of tree children, with the root component being the
KeaTreeRoot component. This change makes the KeaTree component
the root component and the KeaTreeRoot component has been changed
to KeaTreeContents.

This change allows end users to query on KeaTree components and
use the entity to be supplied to the kea_tree_* extensions functions.