Only apply changes to transform if applicable.
The camera::Controller component was updating the transform every
frame even if no changes occurred. Now the transform is only
updated if the velocity length is large enough.

The tools camera transform is only updated when the main camera
tranform is changed.

The Transform widget's scale was being applied every frame. This
is now changed to only be applied when a drag event on the widget
occurs.