Settings module and GUI.
The 'settings' module will contain various settings used throughout the app
and project. This module centralizes all of the settings objects
and logic. The SettingsResource can be used to quickly read/write
settings from anywhere in the code base.

The GUI contains a Settings panel that is displayed through a button in
the Project panel.