Refactored 'shapes' module.
Previously, the module was using a VertexBufferBuilder object,
which also contained some shape building functionality. This
refactor now consolidates all of this functionality into the
'shapes' module.

This change also deprecates the VertexBufferBuilder object.
The 'shapes' module has its own Builder object, mainly to
deal with updating the index buffer with the generated
points and indices for a shape.